Complete mini-spotify

+++ src/model/Music.java
@@ -1,7 +1,7 @@
 package model;
 
 public class Music extends Media {
-    public Music(String title, String creator, int duration, Genre genre) {
-        super(title, creator, duration, genre);
+    public Music(String title, String creator, double duration, Genre genre) {
+        super(title, creator, (int) duration, genre);
     }
 }
